cleaning up planargraphics and image viewer; solving issues with planargraphics (editin moving etc)
